Replace function fromDto from class UserBuilder in dto classes
