+++ src/main/java/com/alevel/trucking/dto/AddressDto.java
@@ -7,14 +7,12 @@ import com.alevel.trucking.model.address.Street;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Positive;
-import javax.validation.constraints.PositiveOrZero;
 import java.io.Serializable;
-
 
 public class AddressDto implements Serializable {
 
     @NotBlank
-    private String  city;
+    private String city;
 
     @NotBlank
     private String street;
