+++ src/main/java/com/alevel/trucking/dto/DriverRegistrationForm.java
@@ -1,8 +1,8 @@
 package com.alevel.trucking.dto;
 
+import com.alevel.trucking.model.person.driver.Driver;
 import com.fasterxml.jackson.annotation.JsonFormat;
 
-import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.Past;
 import javax.validation.constraints.PastOrPresent;
@@ -20,6 +20,25 @@ public class DriverRegistrationForm extends CustomerRegistrationForm {
 
     @NotEmpty
     private DriverLicenseDto driverLicense;
+
+    public static Driver fromDto(DriverRegistrationForm driverRegistrationForm) {
+        return Driver.managerBuilder()
+                .username(driverRegistrationForm.getUsername())
+                .email(driverRegistrationForm.getEmail())
+                .password(driverRegistrationForm.getPassword())
+                .firstName(driverRegistrationForm.getFirstName())
+                .secondName(driverRegistrationForm.getSecondName())
+                .lastName(driverRegistrationForm.getLastName())
+                .phone(driverRegistrationForm.getPhone())
+                .isAccountNonExpired(true)
+                .isAccountNonLocked(true)
+                .isCredentialsNonExpired(true)
+                .isEnabled(true)
+                .birthday(driverRegistrationForm.getBirthday())
+                .startWork(driverRegistrationForm.getStartWork())
+                .driverLicense(DriverLicenseDto.fromDto(driverRegistrationForm.getDriverLicense()))
+                .build();
+    }
 
     public Date getBirthday() {
         return birthday;
