+++ src/main/java/com/alevel/trucking/model/address/Address.java
@@ -1,7 +1,10 @@
 package com.alevel.trucking.model.address;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
