+++ src/main/java/com/alevel/trucking/model/person/driver/Driver.java
@@ -10,7 +10,6 @@ import lombok.*;
 import javax.persistence.*;
 import java.util.Date;
 import java.util.HashSet;
-import java.util.List;
 import java.util.Set;
 
 @Entity
