+++ src/main/java/com/alevel/trucking/model/person/driver/DriverStatus.java
@@ -5,4 +5,9 @@ import java.io.Serializable;
 public enum DriverStatus implements Serializable {
     IN_ROUTE,
     IN_BOX;
+
+    @Override
+    public String toString() {
+        return name();
+    }
 }
