+++ src/main/java/com/alevel/trucking/model/person/manager/Manager.java
@@ -35,20 +35,20 @@ public class Manager extends User {
 
     @Builder(builderMethodName = "managerBuilder")
     public Manager(Long id,
-                    Date startWork,
-                    Date birthday,
-                    String password,
-                    String username,
-                    String email,
-                    String firstName,
-                    String secondName,
-                    String lastName,
-                    String phone,
-                    Set<Role> roles,
-                    boolean isAccountNonExpired,
-                    boolean isAccountNonLocked,
-                    boolean isCredentialsNonExpired,
-                    boolean isEnabled) {
+                   Date startWork,
+                   Date birthday,
+                   String password,
+                   String username,
+                   String email,
+                   String firstName,
+                   String secondName,
+                   String lastName,
+                   String phone,
+                   Set<Role> roles,
+                   boolean isAccountNonExpired,
+                   boolean isAccountNonLocked,
+                   boolean isCredentialsNonExpired,
+                   boolean isEnabled) {
         super(id,
                 password,
                 username,
