+++ src/main/java/com/alevel/trucking/model/transport/TransportStatus.java
@@ -3,5 +3,10 @@ package com.alevel.trucking.model.transport;
 public enum TransportStatus {
     IN_BOX,
     IN_ROUTE,
-    IN_REPEIR
+    IN_REPEIR;
+
+    @Override
+    public String toString() {
+        return name();
+    }
 }
