+++ src/main/java/com/alevel/trucking/repository/OrderRepository.java
@@ -17,5 +17,4 @@ public interface OrderRepository extends JpaRepository<Order, Long> {
 
     List<Order> findByStatus(OrderStatus status);
 
-
 }
