+++ src/main/java/com/alevel/trucking/repository/UserRepository.java
@@ -4,8 +4,6 @@ import com.alevel.trucking.model.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.Optional;
-
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
