+++ src/main/java/com/alevel/trucking/service/customer/CustomerService.java
@@ -3,7 +3,6 @@ package com.alevel.trucking.service.customer;
 import com.alevel.trucking.model.person.customer.Customer;
 
 import java.util.List;
-import java.util.Optional;
 
 public interface CustomerService {
 
@@ -11,7 +10,7 @@ public interface CustomerService {
 
     Customer findByUsername(String username);
 
-    List<Customer>  getAllCustomer();
+    List<Customer> getAllCustomer();
 
     Customer getCustomerById(Long id);
