+++ src/main/java/com/alevel/trucking/service/driver/implementation/DriverServiceImplementation.java
@@ -14,7 +14,6 @@ import com.alevel.trucking.service.driver.DriverService;
 import com.alevel.trucking.service.order.OrderService;
 import com.alevel.trucking.service.user.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
